pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scraper-ai"
version = "0.1.0"
description = "Prompt-driven web scraping agent powered by AI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"

dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "tenacity>=8.2",
]

[project.optional-dependencies]
openai = ["openai>=1.30"]
anthropic = ["anthropic>=0.30"]
gemini = ["google-genai>=1.0"]
all = [
    "openai>=1.30",
    "anthropic>=0.30",
    "google-genai>=1.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
scraper-ai = "scraper_ai.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by line-length setting)
]

[tool.ruff.lint.isort]
known-first-party = ["scraper_ai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
